Criação das tabelas e mapeamento de algumas entidades

+++ projetomonstro-api/src/main/java/com/reaper/projetomonstro/api/repository/GrupoMuscularRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.reaper.projetomonstro.api.model.GrupoMuscular;
 
-public interface GrupoMuscularRepository extends JpaRepository<GrupoMuscular, Integer> {
+public interface GrupoMuscularRepository extends JpaRepository<GrupoMuscular, Short> {
 
 }
